requirements.txt
aiortc
beautifulsoup4
DeepImageSearch
DeepImageSearch
dill
discord_webhook
easyocr
faiss_cpu
ffmpeg_python
Flask
Flask_SocketIO
gevent
ImageHash
imageio_ffmpeg
matplotlib
numpy
opencv-python
Pillow
Requests
setuptools
scikit-image
table2ascii
timm
torch
torchvision
tqdm
websockets
cv2-enumerate-cameras